Add auto-format on save
This is largely taken from the great fatih/go-vim vim plugin. This is a
quick go at isolating the required functionality for `zig fmt` to work
in place of `gofmt`.

You will need to specify the binary path of the zig compiler for this to
work in your zig config. e.g. `let g:zig_bin_path = '/usr/bin'`.