change how the buffer content is replaced
This prevents a bad behaviour when using zls with neovim.

Before, the content of the buffer was replaced without being deleted
first: in neovim this causes a callback "on_lines" to be called _for
each line_ being replaced.

This "on_lines" callback is used to notify LSP servers of changes to the
document, so everytime a 2000 lines file is saved neovim would trigger
2000 calls to zls.

When editing a 2000 lines file this makes neovim hang for a couple of
seconds when saving the file, as well as generating 130MiB of LSP debug
logs.

This commit fixes this by completely deleting the buffer before setting
its content.