Append g:zig_std_dir to &path
Using the local value of path means that if the user has not set a local
value for path, the value of 'path' will just be 'g:zig_std_dir'. By
using '&path' the current value of 'path' is used (i.e. if no local
value is set the global one is used). For example, this means that if
'path' is at its default value of '.,/usr/include/,,' (which is not
local) before loading the ftplugin, before this change the default value
of path would be completely replaced, making gf not find files files
relative to the current file or directtory. With this change, gf will
find imported files relative to the current file and directory by
default.

Appending 'g:zig_std_dir' rather than prepending means that zig files
with the same name as ones in the standard library should be resolved to
the one the project one is working on rather than the one in the
standard library.