remove the zigVarDecl group which maps to Function
The previous mapping made no sense. The mapping of threadlocal and
comptime will be changed in future commits.

As a reference I used the builtin syntax highlighting from neovim for
the following languages that have similar keywords to Zig's `var` and
`const`:

- Rust: `let`
- Go: `var`
- Javascript: `var`,`let`, `const`

They all map their keywords to the Keyword syntax highlighting group.